+++ CHANGELOG.md
@@ -8,6 +8,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 * Support for OPML imports and exports
 
+### Fixed
+* Secure cookies are no longer sent when debug mode is enabled
+
 ## [0.1.0] - 2022-04-14
 Initial release